Support multi value server flags
This change introduces support for multi-value flags including the
command line parameter for overriding flags. The format is:
-Dstudio.server.flags.enabled.override=analytics/surveys/followup/$index
where index is the position of the target FlagValue in the config.

The switch from single value to multi value format flags is gated by
the constant SUPPORTS_MULTI_VALUE which is set to false. However, the
underlying data structure of server flags is changed with this CL.

Studio stats related changes will be made in a different change.

Test: new tests
Bug: 380309577

Change-Id: I8b90c8af01366f8c19a756bab5e8e0cc59406e87