Fix paginated table sorting by string rather than value
It was found that when sorting a column in the pagintated table
view in Profilers, it was sorting based on the values string value,
rather than its actual value. This was because the comparator for
each column was not being set, so it kept defaulting to using a
string comparator. This change rectifies this.

The call to allRowsChanged() was also removed as it was proven to
be doing nothing.

Fixes: 380490476
Test: manual
Change-Id: I10f6a16c40789647b654a61f284d31d8b127158e