Migrate 'Configure SDK' wizard
If you go to 'Languages & Frameworks -> Android SDK' in 'Settings' and
click on 'Edit' next to 'Android SDK Location', a wizard is launched
which allows you to change the path. It re-uses the steps from the
welcome wizard.

Logic has been added to use the new wizard when `NPW_FIRST_RUN_WIZARD`
is `true`.

An issue in the existing implementation has also been fixed - progress
is now correctly reported when the progress step is shown (the modality
state was missing).

Bug: 311256502
Test: Manual
Change-Id: I308345cad83263af3a83d9a34eaa71ba57f216b1