Refactor to move Ask Gemini query building out of project-system-gradle
The APIs needed for sending error source files as context are only
available in the internal APIs in aiplugin.android module. They will be  accessed through extension points implemented in the module. Also, the prompt queries constructed for gradle errors are expected to change frequently. By separating out this logic from project-system-gradle,the queries can be modified independently.

This CL:
1. Introduces a new extension point 'StudioBotQuickFixProvider' that
   is declared in project-system-gradle module. This is used for placing
   chat queries for 'Ask Gemini' use-cases. This will be implemented in
   aiplugin.android module.

2. Removes test code that tests query formatting. It will be moved to
   the implementation side of 'StudioBotQuickFixProvider'.

Bug: 316057751
Test: existing
Change-Id: I826f7638c85fac4df8713b43f77f804bfdc7ae6b