[Dsl,.kts] Improve support for domain object configurators
Including "configure" in the list of configurators was a mistake: that
configures the container, not the individual object.  Remove it, but
include "named" which should have been present.

Move the utility function for determining whether a given method is a
configurator to shared parser utils, so that we can implement support
for these configurators for Groovy scripts.

Add a test (currently KotlinScript only) for all configurator methods.
Also include in it something which looks syntactically like a
configurator but isn't, to check that we don't include it.

Also add a convenience test method which should lead to less
confusion (so that `assertEquals(string, model)` doesn't fail with the
message "expected:.one; actual:.one".

Bug: none filed
Test: added
Change-Id: I96c08f5b680244178d55b7d5de4580337fe822c8