Migrate Android plugin 'SDK Setup' wizard
This wizard is displayed for IntelliJ IDEA apps that have the Android
plugin configured. It can be triggered by clicking 'New Project' from
the 'Welcome screen', clicking 'Android' in the 'Generators' options on
the left and then clicking the 'Install SDK' button.

Note: the button will only be show if no sdk has been previously
configured, i.e. it is `null`.

Bug: 311256502
Test: Manual
Change-Id: If2d5a6faaf48605de29f492df71e4b6fe52993e0