Fix new welcome wizard border issue
The scroll views are showing with a border in the new wizard. They are
overridden to `EMPTY` in the old wizard, so let's replicate that here.

You can force the new wizard to launch when running Android Studio
locally by setting the `NPW_FIRST_RUN_WIZARD` and `NPW_FIRST_RUN_SHOW`
flags to `true` in `StudioFlags.java`.

Bug: 311256502
Test: Manual
Change-Id: Ied889f1dd7740c7b10fac5fc579f14972554519b