Update SDK category node 'selected' logic
A category will currently show as 'selected' when none of its children
are selected. This is unintuitive (see the screen recording), and does
not reflect the `toggle` logic, which updates all children to the same
value as the parent. Let's update it to show as 'selected' when all of
its children are selected, and 'unselected' otherwise.

Screen recordings (ignore the duplicate platform entries, I added them
manually to highlight the issue in the recordings):
* Before - http://screencast/cast/NTA5MTU5Mzc2NDczMjkyOHxiZjVlZWY0My03MQ
* After - http://screencast/cast/NTI3MDUxNjkzNDkwMTc2MHw0ZjhiN2Y2YS0yNw

Bug: 311256502
Test: Manual
Change-Id: Ia28c428458446aafdbcc41f9dc3aaee25ee264c3