Fix wizard crash when in INSTALL_HANDOFF mode
The old wizard will crash when running in INSTALL_HANDOFF mode
(installed with Windows installer) when the installer has already
installed an SDK.

Note: When testing on a Windows Cloudtop VM, the Windows installer
downloaded from https://developer.android.com/studio doesn't have an
option to configure the SDK. It also generates a firstrun.data file with
an empty version, which means the `INSTALL_HANDOFF` mode will never
actually be triggered.

So as it stands, it's unlikely this crash is actually being triggered
for any users. Let's fix though, in case the installer is re-configured
in the future.

Bug: 311256502
Test: WelcomeScreenWizardTest
Change-Id: I6242553d12d8ab6a0881a1abbb55de933f0e0977