Remove start point dropdown conditional disablement + auto-select
One feature in the Profiler's task starting point dropdown was to
auto-select the other task starting point dropdown option if the
other was disabled. The other related feature the former feature
relied on was conditional disabling the dropdown option if it was
not possible with the selected configuration.

However, after much user-feedback, it was clear that auto-select
and the conditional disablement of the dropdown options was
causing more frustration and confusion.

Thus, in this change, the conditional disablement and auto-select
features were removed.

Fixes: 364466904
Fixes: 376044038
Test: updated
Change-Id: I61cd9266f0a600572049b3bcce22469bebe8d494