Show suggested and other app both in Vitals popup
Vitals app selection popup only shows a section if it has at least one
element inside it. With this change, it shows both sections, suggested
and other, with an appropriate message if the section is empty.
This change also adds a banner when the user has no access to any apps
in Vitals

Fixes: 295897640
Test: Updated
Change-Id: Idf4a303a9388843685cdcff7ef55069151341887