Fix invalidate on cancellation race condition
The invokeOnCompletion can be run concurrently to a new refresh request.
This led to some rare race conditions in CommonPreviewRepresentation
where the invalidate would be set to true in the invokeOnCompletion but
after the new refresh had already checked if an invalidation needed to
happen. That problem was fixed in I472e1cf50b05890cd524.This CL fixes
the same potential issue for Compose previews.

Bug: N/A
Test: ComposePreviewRepresentationTest x1000
Change-Id: I63e9c7c0331c2c62bea4d6f46f20292ea77b6924