Add suggest a fix core functionality
This change introduces an extract phrase in the insight prompt to Gemini
asking it to return a single file (only single file supported for now)
in which the fix is performed. When users click on the "Suggest a fix"
button, it calls the EP implementation FixSuggester in the android
aiplugin module which calls the TransformService, providing it with the
file and the insight, and if a fix is found, brings up a diff view in
which users can accept the changes to their code.

In AQI, the insight bottom toolbar actions have all been moved to the
right side, making room for the new "Suggest a fix" button at the bottom
left.

Parts that need more work:
- UI polish in the bottom right of insights panel
- Caching is not optimal after the introduction of the prompt for code
context
- Integration with studio labs.

Test: new and existing tests.
Bug: 413661039
Change-Id: Ic74d49cfc29cfbd82e30d642bb102c421b505b05