More careful ApkProjectSystem instantiation
Since in principle we might open projects which have a persisted
project system provider ID of the APK project system in an Android
Studio with the flag turned off, make sure to respect the flag in the
ApkProjectSystemProvider factory as well as the test for
applicability.  (This is to address a scenario that is vanishingly
unlikely to happen in production).

Bug: 325936093
Test: existing
Change-Id: I16bad35dcf242e3a5282d4266a4467f913ed021b