Improve how we disable AS analytics in IJ IDEA
Previously we disabled AS analytics inside a
ActionConfigurationCustomizer, but it is better to do this inside a
ApplicationInitializedListener which runs at a better defined point in
the IDE startup sequence.

This change was inspired by JetBrains/android commit f50b03b5cb.

A unit test is not feasible (since it depends on AS running with a
different platform prefix), but this behavior is implicity tested by
JetBrains.

Change-Id: Ic5d56208770962df07b9953e54fa6e5d8388fc6f