K2 LiveEdit: find module using ModuleUtilCore.findModuleForFile()
This change was recommended from JetBrains. They said:

"In many cases, PsiFile objects might become invalidated. In our case,
for example, this happens quite frequently when just editing a code in
a KMP project's common sources due to background re-parsing for
cross-referencing from Swift. The implication of such invalidation is
that extension function KtFile.module starts returning null and the
compilation (of a preview) immediately finishes with an internal error.
So, the idea here is to use a less-restrictive method to provide the
file's module."

This is the related upstream change:
https://github.com/JetBrains/android/commit/c6262f02dcd4f1f813e4820e995861251c8fff54

Test: N/A
Bug: 383564069
Change-Id: I61138af0d936b2579beff67a6712c63731b43ba2