Properly handle all Compose group APIs
Live Edit was grabbing the wrong stack value when handling Compose group
APIs that accept an additional parameter beyond just the group key
(startReusableGroup and startMovableGroup).

This CL properly handles those group methods, as well as widens the
scope of JVM bytecode operations that we consider to "add an int
constant to the stack".

Test coverage is added to ensure we can properly parse
startRestartableGroup, startReplaceableGroup, startReusableGroup, and
startMovableGroup (startReplaceGroup is still too hard to generate in a
test context).

Fixes: 383382470
Test: ComposeAnalysisTest
Change-Id: I6e26addee4275b09d81e89594ed2f9b8ec87b522