Replace runWhenSmartAndSyncedOnEdt with runWhenSmartAndSynced
in RenderingBuildStatusManager.kt

Using runWhenSmartAndSyncedOnEdt and then immediately launching a
coroutine on workerThread within the callback was redundant and added
unnecessary complexity to the EDT scheduling.

This change exposed other race conditions, changes for them:
1. Fix premature call to updateAnimationPanelVisibility() during
initialization, by moving hasRenderedAtLeastOnce to the first line
2. Remove the onReady callback from RenderingBuildStatusManager as it
was being invoked during the initialization of
ComposePreviewRepresentation. We can invoke that logic after
initialization of ComposePreviewRepresentation

Fixes: 384037327
Test: RenderingBuildStatusManagerTest
Change-Id: I98d7dad3557ba521c1fded9cdf8b8fe9039da841