K2: Drop document invalidation from ComposePreviewRepresentationTest
The document invalidation right after creating the virtual file
generates a document with null virtual file, and it is cached. When we
have a document for the virtual file later, updating PSI updates the
document for the virtual file. However, we still have cached the
document for the null virtual file, and it causes the inconsistency
from the one for the virtual file.

Bug: 381432038
Test: K2 ComposePreviewRepresentationTest.testRerunUiCheckAction
Change-Id: Ie29fc25538b60f91e796d3ee8c32f90345c5609b