[Dsl] Refine handling of GRADLE_PROPERTY elements
In both Groovy and Kotlin Script files, accept parsing of the
assignment form.  For Groovy this has been allowed since time
immemorial; in Kotlin Script it is only supported since Gradle 8.2, so
even though users probably prefer to have it written out in assignment
form, be conservative and emit assignments for GRADLE_PROPERTY only in
Groovy.  (Subsequent work to allow us to track the current version of
Gradle used in the project might allow us to behave more nicely in
Kotlin Script).

Bug: 368629093
Test: adjusted and added
Change-Id: I431f9d8e39d13b465beb8616e58897ec19fcb14d