[Dsl] Support defaultLibrariesExtensionName.set("libsCustom")
Before, a custom alias for the default version catalog (`gradle/libs
.versions.toml`) was considered only if it was set via assignment in a
settings file:

```
dependencyResolutionManagement {
    defaultLibrariesExtensionName = "libsCustom"
}
```
This commit:
1) Makes possible processing both ASSIGNMENT and SET_METHOD syntax for
GRADLE_PROPERTY as discussed in the Gerrit review:
https://android-review.googlesource
.com/c/platform/tools/adt/idea/+/3401659

2) Processes `defaultLibrariesExtensionName` in
`dependencyResolutionManagement` as GRADLE_PROPERTY, not VAL.
This property specifies a name for a version catalog with the default
location (gradle/libs.versions.toml).
This change adds support for specifying this property not only via
assignment (as it was before) but also via property setter.

fixes IDEA-303093

Fixes: 383237356
Test: added
Change-Id: I027ea314f96475bbfe5189cc778c68581e508f0e