Remove startup metrics for power save mode
This code triggers initialization of the TogglePowerSave action too
early during IDE startup, leading to the error in b/379126565.

These metrics are not expected to be used anyway, because we are no
longer investing in power-save / essential-highlighting modes.

We'll continue to receive metrics in the event where the user explicitly
enables or disables power save mode (as part of our general action
invocation metrics).

Bug: 379126565
Change-Id: I706f80b19eedfb15ab2c145d6b7a2710bdbe64df