Refactor AndroidStudioWelcomeScreenProvider and add tests
Extracted most of the logic to a new AndroidStudioWelcomeScreenService
class to make it easier to test.

Also removed logic that was causing the `NPW_FIRST_RUN_WIZARD` flag to
act in the same way as the `NPW_FIRST_RUN_SHOW`, i.e. showing the
wizard every time Studio is launched.

Bug: 311256502
Test: AndroidStudioWelcomeScreenProviderTest
Change-Id: I7aa7a30197fd6eb7234f75962328480e4fde1726