Improve RootsChangedDumbModeTask to handle project disposal
..and task cancellation

1.  Added `indicator.checkCanceled()` to `RootsChangedDumbModeTask
.performInDumbMode` before each iteration of the loop that processes
modules. This allows the task to be cancelled more promptly if the
`ProgressIndicator` is cancelled
2.  Added `project.isDisposed` check to `RootsChangedDumbModeTask
.performInDumbMode` after `indicator.checkCanceled()` to abort task if
project is disposed.

Previously, the task only checked for project disposal at the beginning
of `performInDumbMode`. If the project was disposed of during the loop's
execution, the task could still attempt to access disposed resources,
leading to errors and memory leaks.

Bug: 384688535
Test: n/a
Change-Id: I476892c7dfb5534b3eaeff4fc0ca8f71807917cd