Support K2 compose preview for multiple KT files
The existing K2 compose preview compiles KT files one by one without
duplication check. As a result, when two KT files have dependency on
each other (let's say they are A.kt and B.kt), we compile each of them
twice for the target A.kt and the target B.kt. As a result, the same
class files will be generated twice. The compose preview tries to write
them to the file system, which results in FileAlreadyExistException. To
avoid this type of exceptions, this commit lets K2 compose preview avoid
compiling duplicate KT files and writing duplicate class files.

Bug: 352403444
Test: FastPreviewManagerGradleTest.testMultipleFilesCompileSuccessfully
Change-Id: Ifa3e7c6e8daf07a71f1339a5b17fcd0f517fcfb9