Set a foster parent for InformationPopup's content to
... populate the popup's actions' DataContext with additional data.

This CL registers a foster parent for InformationPopup's contents. The
parent is the owner that triggered the popup to show. This ensures that
the DataContext is also populated by using the owner's data providers
and allows actions within the popup to query for data that is not
directly related to the popup, such as PlatformCoreDataKeys.

This replaces the previous mechanism of registering a DataProvider to
the popup's component that would delegate to the owner's DataContext.
This was a problem for a couple of reasons:
1. the DataContext that was being delegated to could be stale
2. when creating a new DataContext, we would query the existing
   DataContext. This is not allowed and caused a crash

Bug: 385329136
Bug: 384098143
Test: manually tested by checking the data context in the
 popup's actions is populated with PlatformCoreDataKeys elements. If the
 foster parent is not set, those keys should be absent.
Change-Id: I4933eb74dcd5d2dcc488a3e3edad12b41e045de1