Rewrite CommonIssueNotificationActionTest to use SimpleDataContext
The new view of DataContext objects is that they have snapshot-like
behaviour: essentially, early-binding rather than late-binding, taking
the values at construction rather than looking them up on dereference.

This means that we need to be careful to construct a new data context
in the test each time the state changes, rather than relying on the
old data context to look up the new value.  Each state change in the
tests is therefore triggered by generating a new event, constructing a
new DataContext object at that point through the custom get() for the
dataContext property.

Bug: 382112784
Test: existing (test-only change)
Change-Id: I7ae2d31b373e9940f192cc24f020e97bf5939781