[AUA] Extract change in default from block property processor
AbstractBlockPropertyUnlessNoOpProcessor blocks the upgrade in these two
conditions:

- When the property is used and the upgrade goes after propertyRemovedVersion
- When the upgrade comes from a version before defaultChangedVersion

This change extracts the part that blocks when a the upgrade starts
before the change in default into AbstractBlockPropertyWithPreviousDefaultProcessor
so the part that blocks when the property is used can be extended in new
processors for properties that do not have change in default.

There is no change in functionality, but this will allow to create new
processors.

Bug: 388889450
Test: BlockRenderScriptPropertyPresentRefactoringProcessorTest
Test: BlockAidlPropertyPresentRefactoringProcessorTest
Change-Id: Ibe8cc0fc0d1269286005649899d86bed905d10d2