Fix CreateTypedResourceFileAction
When migrating the code to Kotlin, references to the
ResourceFolderType#getName method were moved to the Enum#name method
instead because of the getter naming conflict.
Enum#name returns the enum name as is, so for folder types this will be
uppercase while ResourceFolderType#getName returns the folder name
(lowercase).

Test: CreateTypedResourceFileActionTest
Fixes: 385766062
Change-Id: I82204201b7158314b7fca7ff5bade063ad6dcfd1