Plumb metrics into UrlFileCache
This change introduces the ability to get some metrics from
UrlFileCache. It also makes some parts of that class more
explicitly tested.

Specifically, it allows callers of the "get" method to specify
a consumer of a new FetchStats data class, which will be called
exactly once when the "get" operation finishes, whether or not it
succeeds. The value of the FetchStats object can then be used to
report metrics about what happened.

Bug: 319684997
Test: UrlFileCacheTest.kt
Change-Id: I2ee1bb88ce72101a2047da996974bb4518fa7e22