Restore check for AndroidModule existence
Some recent changes to the class got rid of the check to ensure that the
tests are being invoked from within a AndroidTest module, which caused
b/389733593. This was not caught by tests because it will not happen all
the time unfortunately. This is only reproducible when the project does
not have androidTest srcSet configured, and the checks we have above in
this function will not return false (as expected) because the project
will not have any androidTest resources to check the file intersection
with them.

By restoring this check, we make sure that we are trying to run the
instrumented tests from a AndroidTest module, which will correctly
return false as expected for this corner case.

Fixes: 389733593
Test: Added
Change-Id: I8311e6426fa8ec06deb887efdb24c2f9aaa139de