Delete a test
This test is not desperately well motivated: it seems to be testing
that we fail to detect a relevant change because some part of the
GradleFiles infrastructure is not active.  This is more of a "how is
this implemented" test than anything else, and gets in the way of
reworking GradleFiles.

Bug: 345518607
Test: remaining tests
Change-Id: I02e33e6bc8be768cd6a4ed0c84a6c2cec1d7f1b7