Use module pointers in linked modules group
I think it was considered safe to hold direct modules references in
this group because it can only change all-together as a result of sync.
But apparently modules can be reloaded by JpsProjectModelSynchronizer
from file changes in .idea folder without sync (see b/387499844).
This leads to a situation when we hold a disposed module and usually
not checking if it is disposed, which leaded to 'module already
disposed' crashes after JpsProjectModelSynchronizer activity.

Using module pointer instead of holding module directly would help
to avoid such crashes in case of unexpected modules renaming or
deletion.

However, the problem of suddenly lost GradleFacet and NdkFacet
still remains, as these are only attached to holder modules and thus
are not accessible with holder module disappearance. Gradle sync should
restore normal order of things in this situation.

Bug: 387499844
Test: N/A
Change-Id: Idcda4929c1a32a596a7fc856c7895f3d32da5b3c