Disable KotlinGradleTomlVersionCatalogReferencesSearcher
With K2 when renaming a version catalog plugin name the
KotlinGradleTomlVersionCatalogReferencesSearcher no longer gets
ignored since BaseKotlinUastResolveProviderService implements
FirIdeaKotlinUastResolveProviderService instead of
IdeaKotlinUastResolveProviderService being this expected, however,
this change is causing PsiElement to be able to be resolved and
as consequence the reference searcher will be taken into account.

All of this is conflicting with studio implementation of reference
searcher called GradleKtsVersionCatalogReferencesSearcher resulting
on having both enabled and causing randomness on the result. With the
addition that platform one doesn't support properly the above operation
resulting on wrong version catalog reference.

Bug: 381140436
Test: Existing
Change-Id: I66a6d8cd20dfc7e1e0058b443170e7adc3c60b89