Build workflows cleanup in tools/adt/idea (Part 1)
1. Adds the new method to obtain a build invoker from a build system.
2. Adds the new methods to invoke blaze/bazel commands via the invokers.
3. Minor changes to the BazelExitCodeException class.

This CL does not affect the users. The AbstractBuildInvoker1 class is added temporarily (to make the CL compile), and it will replace AbstractBuildInvoker in the subsequent CL.

The subsequent CL will replace all usages of AbstractBuildInvoker with
the new one, and get rid of existing invokers, command runners, and build result helpers.

The lint warnings about java_library are unrelated to the change and
seem safe to ignore.

Bug:374906681

Test: n/a
Change-Id: Ia5a6ba4490e4b32cd8c16c572b3a1feae7f5dcac