Remove build system check in NewVectorAssetAction
The check, for whether PNG generation is supported, and subsequent
check against the minSdk version, is substantially incomplete, eliding
as it does any mention of the support library's varying compatibility
support for vector drawables.

Given this, the check is of limited utility, and (worse) actively
prevents the user from doing what they need to in cases where their
action is completely valid.  Remove the check, which also makes the
question of whether the build system automatically generates PNGs for
vector drawables irrelevant.

Bug: none filed
Test: existing
Change-Id: I11a3df3b57df714f5f4e71ced88bae9a9ad7e0d2