Add 'LOAD Alignment' APK Analyzer column
This adds a new column called 'Load Alignment' which indicates the minimum ELF LOAD segment.
It also renames the existing 'Alignment' column to 'Zip Alignment' to disambiguate the two.
This feature is controlled by a server flag and is off by default.

Screenshots available in the in-progress bug bash instructions: go/as-16kb-bugbash

Bug: 382091362
Test: Manually verified with and without server flag enabled
Change-Id: Id527d4b9eba299dffa11a1851d84c9b137df00b0