Stop checking warnings in import tests
AndroidMavenImportFixTest.testSuggestedImport_kotlinFile is failing in
K1 mode because it returns a warning that K2 does not.

None of the test cases in this test actually care about any warnings;
they only care about an error. Turning off verification of warnings in
all tests, so that the difference between K1 and K2 no longer matters.

Fixes: 393599143
Test: AndroidMavenImportFixTest
Change-Id: I23b445c0c58b1f0c1e62a5990ded74361325f2b6