Don't force-close On Detach
In `DebugSessionStarter.attachDebuggerToStartedProcess()` we add a
`ProcessListener` that ends up calling `adb am force-stop`. This happens
even if we `detach` from the debugged process.

In `ProcessHandler`, both `destroyProcess()` and `detachProcess` fire
a `processWillTerminate()`. The difference is the `willBeDestroyed`
arg is `true` for `destroy` and `false` for `detach`.

The `willBeDestroyed` is surfaced on the
`ProcessListener.processWillTerminate()` method so we can avoid calling
`force-stop` when detach while still calling if for `destroy`

This is needed for this cl: ag/31574967

Bug: b/393393475
Test: Added
Change-Id: Ia5a56ee2cb9693dc07be0149003c33fa38ec65c9