Turn on javadoc/source multi variant support experimental flag
A boolean is added to ModelBuilderParameter to instruct agp whether to
obtain the addtional artifacts and pass them through the model to studio.

When new studio work with slightly older agp version(e.g. 8.2), if the
user enables the agp experimental flag, there will be a duplicate work
of fetching on both agp and studio side. However, this won't be an issue
for most users because the agp flag is in an experimental state.

The IdeLibrary is updated to reflect whow we model source jars today.

Bug: 271601839
Test: existing
Change-Id: I397541866e81d039fab31bf4e84e64e9010e4c95