Copy Gradle Daemon log file instead of moving it
As this causes issues on Windows where the file lock is still being held when reaching this stage.
Making sure we close the daemons before moving the file does not work neither (checked that no Gradle Daemon connection is open at this stage), so in order to avoid this issue, we can just copy the file instead.
There are other places where we move files in our tests, but I haven't seen file access issues about them.

Bug: 378648219
Test: N/A
Change-Id: Ic12812e1c3a83288f5d2a71322d7ff64979ec1d0