[Sync] New model for declared / registered dependencies
Various parts of the IDE want to be able to query the project system
about declared (registered) dependencies, in order to know how or
whether to make changes to build configuration files.  In the current
Gradle implementation, we historically parsed build files directly,
before turning that off (because parsing build files is an
intrinsically slow operation) in favour of falling back to resolved
dependencies on the classpath -- but resolved dependencies might have
been transitively included rather than explicitly specified, and do
not reflect the user's declaration but have versions resulting from
dependency resolution.

This change introduces a new injected model builder to query Gradle
directly at sync-time for the dependencies registered in specific
configurations.  We collect information for five specific
configurations, because of the eventual desire to be able to answer
the dual questions of the registerDependency operation, which allows
for registering debugImplementation and annotationProcessor
dependencies as well as implementation.

Bug: 129297171
Test: enabled in next change
Change-Id: Ib771faadac7e2e5d7c52629a37eb039befc0c809