[DCL] Don't spuriously allow Gradle sync seem to fail
Although we catch the UnknownModelException, that's not enough to stop
Gradle from reporting it, so it looks like Gradle sync fails (though
it reports a success.)

Fix by using findModel() rather than getModel(); since findModel()
returning null is an expected case, only LOG that at debug level.
However, errors and incompatibilities (for instance as of this CL,
project isolation) can cause errors in building the
DeclarativeSchemaModel, so preserve the catching of exceptions for now
to allow our project isolation tests to continue to work.

Bug: 394598451
Test: manual
Change-Id: Ib85e705a6ec176bee4e5b6dcf06ae745e3c5eb7b