Add on-device renderer for xr
This change adds basic support for on-device rendering for embedded LI.
There are two renderers available: StudioRendererPanel and
OnDeviceRendererPanel. Both are hosted by RootRendererPanel.

When the LayoutInspectorModel detects XR, the RootRendererPanel switches
the renderer from the studio renderer to the device renderer.

The device renderer sends rendering instructions to the device and
receives input events from the device.

Test: OnDeviceRenderingClientTest
Test: LayoutInspectorManagerTest
Test: OnDeviceRendererPanelTest
Test: StudioRendererPanlTest
Test: SelectedTabStateTest
Bug: 398201661
Change-Id: I8f0b1f4620979a26486f7ed988c295f17f4d392b