Move ExplainSyncOrBuildOutput to aiplugin
Instead of providing this action directly in
AndroidGradleExecutionConsoleManager implement a special group action
where any plugin can register additional actions.
ExplainSyncOrBuildOutput is now registered in plugin.xml and added
to this group.
Additional bonus of a proper action registration is that it can also be
accessed via action search when applicable (when build tree node is
selected).

Test: Adjusted existing tests. ExplainSyncOrBuildOutputIntegrationTest
was changed to check against different issue to make it simpler.
Bug: 395561186

Change-Id: I556db0b9e6c5539816f2a7fa124de9d36207d036