Retrieve the SDK dynamically
The SDK path may change as a result of a prior wizard step that installs
the SDK. A prior fix (I58ed8806506777) worked around this by directly
updating the path of the RepoManagerImpl. This makes the RepoManagerImpl
inconsistent with its parent AndroidSdkHandler.

We can do this more cleanly by fetching the AndroidSdkHandler only when
we enter the wizard step.

A similar change was made in I39439b7194fa5ec to LicenseAgreementStep.

Bug: n/a
Test: n/a
Change-Id: I22d26e3c5f7a85caaefc28efdb3c95e343c61755