Disable lint check while pre-1.8 spec is somewhat supported
We've made a decision to add limited support to the legacy device spec
while Compose 1.8 is not stable for a few months. The limited support
makes sure that using known constants, i.e. Devices's TABLET or PHONE
keep working during this transition.

We successfully applied this behavior in change Ic6dccfc, but we didn't
update the lint inspection we had in place to detect usage of the old
spec. This change disables the lint check, and the plan is to re-enable
it once we drop pre-1.8 spec support completely in the future.

Fixes: 397120285
Test: InspectionsTest
Change-Id: I99459514d38e45313415ced974b581154ae5cfa5