Some UX fixes for mlkit file editor
If the mlModelBinding feature is off in AGP, then there are no source
provider directories, so there is no way that we ever showed the
BuildFeatureOff panel.  Remove the check for being in an mlModel
folder.

Also, when the build feature is off, then the fileEditor window we
create might have no focussed editor.  Under those circumstances, the
editor notifications are never updated, so schedule an update by hand
in the editor constructor.

This does not fix all UX problems with the file editor.

Bug: none filed
Test: manual
Change-Id: I7c84b9573dbe5c77f28a3df6f39d4f855f3e974c