Restore tree selection after tree panel is recreated
The selection in the tree panel can go out of sync with the selection in
the inspector model. This can happen for example if the tree panel is
recreated while the inspector model already has a selection. When the
new instance of the tree panel is created, the listeners are added back.
If the selection listener is executed before the modification listener,
restoring the selection in the tree will fail, because at that point the
tree is still empty.

Test: NA
Bug: NA
Change-Id: I6a7c952cb365f24f14952ad27a8e6b9fcdaf8f1b