Introduce context menu in the Running Devices window
The context menu is disabled by default. It will be enabled when screenshots and screen recordings start supporting non-primary displays.

When the context menu is enabled, it is advertised in a balloon that is shown below the Running Devices toolbar so that the user is not confused by the suddenly disappeared toolbar buttons.

Bug: N/A
Test: existing; more tests will be added when the context menu is enabled
Change-Id: I2d1f6e4b6706dc0d739264f04615c331cf3600c3