Add tests of RegisteringModuleSystem methods
Make sure that the module system returned from a known project system
has a specific subtype, including for the TestProjectSystem, in order
that we can test the details of the return results (where appropriate)
and not just nullity or non-nullity.

While we have both codepaths (the legacy one directly from the
AndroidProjectSystem, and the new on in RegisteringModuleSystem
methods), verify both in DependencyManagementTest.  In other tests,
convert to the new form where the operation is generic; for now, leave
more specific cases alone.

Bug: 279149611
Bug: 279886738
Test: added
Change-Id: I90098b6f749a9bf7519c2fce748d3c57c6e8a71a