Add Horizontal scrolling to component tree
A custom horizontal bar was used to manipulate the look
and feel of each tree node. The width and x-axis of each node
were adjusted according to the horizontal bar's value,
and the horizontal bar's range was updated based on
the width of the widest expanded tree node. And
set component cell renderer opaque status to false and
adjusted the max rendered width.

Bug: 336309065
Test: TreeTableImplTest
Change-Id: Ie96c75ac9c480e8f708c39b66d4802c59bca943d