Move existing icon file check to background thread
When the 'Asset Studio' is first loaded (right click 'res' directory,
then 'New -> Image Asset') there is logic to check whether any of the
icon files that would be generated already exist. If they do then a
warning is shown.

The logic to fetch the paths of all the icons that would be generated
can involve some I/O - when a drawable is configured it needs to parse
the xml output to see if the min api version should be included in the
path. Since this check is currently done on the UI thread, it can result
in the UI freezing while the check is carried out.

To fix this the check has been moved to a background thread.

Bug: 390731020
Test: Manual
Change-Id: Ib994abaac100ef1cc33ceb730ee81f38dcff8e8b