Draw inlay in line center manually
Change the height of the inlay presentation to editor line height and
change the draw coordinates that center the text. This removes the need
to rely on Intellij's line centered inlay. As an added benefit this
allows the cursor to have behavior similar to the one on the file link
next to it.

Fixes: 295392877
Test: Manual
Change-Id: I09ec2f7f402f69a93b72bb5d4b6359454e9db448